+++ mvp/src/main/java/com/mandria/android/mvp/rx/callbacks/OnNext.java
@@ -1,5 +1,7 @@
 package com.mandria.android.mvp.rx.callbacks;
 
+import android.support.annotation.NonNull;
+
 import io.reactivex.functions.BiConsumer;
 
 /**
@@ -7,4 +9,6 @@ import io.reactivex.functions.BiConsumer;
  */
 public interface OnNext<U, V> extends BiConsumer<U, V> {
 
+    @Override
+    void accept(@NonNull U u, V v) throws Exception;
 }