+++ mvp/src/main/java/com/mandria/android/mvp/rx/proxies/AbstractSubscriptionProxy.java
@@ -1,6 +1,5 @@
 package com.mandria.android.mvp.rx.proxies;
 
-import com.mandria.android.mvp.MVPLogger;
 import com.mandria.android.mvp.rx.BoundData;
 import com.mandria.android.mvp.rx.RxView;
 
@@ -8,7 +7,6 @@ import io.reactivex.Notification;
 import io.reactivex.annotations.NonNull;
 import io.reactivex.disposables.CompositeDisposable;
 import io.reactivex.disposables.Disposable;
-import io.reactivex.functions.Action;
 import io.reactivex.functions.BiFunction;
 import io.reactivex.functions.Consumer;
 import io.reactivex.functions.Predicate;
@@ -17,13 +15,6 @@ import io.reactivex.functions.Predicate;
  * An abstract class to manipulate subscription proxies.
  */
 public abstract class AbstractSubscriptionProxy<View, Result> {
-
-    private final String mTag = getClass().getSimpleName();
-
-    /**
-     * Action wrapper to run on original stream termination.
-     */
-    final Action mOnTerminate;
 
     /**
      * Combine latest bi-function to apply.
@@ -47,26 +38,9 @@ public abstract class AbstractSubscriptionProxy<View, Result> {
 
     /**
      * Constructor.
-     *
-     * @param onTerminate Termination action to run.
      */
-    AbstractSubscriptionProxy(final Action onTerminate) {
+    AbstractSubscriptionProxy() {
         mCompositeDisposable = new CompositeDisposable();
-
-        // Wraps onTerminate action to dispose all disposables
-        mOnTerminate = new Action() {
-            @Override
-            public void run() throws Exception {
-                // This method is called after termination consumption
-                // we can dispose all
-                mCompositeDisposable.dispose();
-                try {
-                    onTerminate.run();
-                } catch (Exception e) {
-                    MVPLogger.e(mTag, e.getMessage());
-                }
-            }
-        };
 
         mCombineFunction = new BiFunction<RxView<View>, Notification<Result>, BoundData<View, Result>>() {
             @Override