+++ mvp/src/main/java/com/mandria/android/mvp/rx/proxies/FlowableSubscriptionProxy.java
@@ -1,16 +1,13 @@
 package com.mandria.android.mvp.rx.proxies;
 
-import com.mandria.android.mvp.MVPLogger;
 import com.mandria.android.mvp.rx.BoundData;
 import com.mandria.android.mvp.rx.RxView;
 
 import io.reactivex.BackpressureStrategy;
 import io.reactivex.Flowable;
-import io.reactivex.Notification;
 import io.reactivex.Observable;
 import io.reactivex.annotations.NonNull;
 import io.reactivex.disposables.Disposable;
-import io.reactivex.functions.Action;
 import io.reactivex.functions.Consumer;
 import io.reactivex.processors.ReplayProcessor;
 import io.reactivex.subscribers.DisposableSubscriber;
@@ -21,8 +18,6 @@ import io.reactivex.subscribers.DisposableSubscriber;
  */
 public class FlowableSubscriptionProxy<View, Result> extends AbstractSubscriptionProxy<View, Result> {
 
-    private final String mTag = getClass().getSimpleName();
-
     private final DisposableSubscriber<Result> mReplayDisposable;
 
     private Flowable<BoundData<View, Result>> mFlowable;
@@ -32,10 +27,9 @@ public class FlowableSubscriptionProxy<View, Result> extends AbstractSubscriptio
      *
      * @param flowable    Original flowable.
      * @param view        Observable that emits the view.
-     * @param onTerminate Action to perform when the ReplaySubject will terminate.
      */
-    public FlowableSubscriptionProxy(Flowable<Result> flowable, Observable<RxView<View>> view, final Action onTerminate) {
-        super(onTerminate);
+    public FlowableSubscriptionProxy(Flowable<Result> flowable, Observable<RxView<View>> view) {
+        super();
 
         // Creates a replay subject which will subscribe to the flowable.
         final ReplayProcessor<Result> replaySubject = ReplayProcessor.create();
@@ -67,20 +61,7 @@ public class FlowableSubscriptionProxy<View, Result> extends AbstractSubscriptio
                         replaySubject.materialize(),
                         mCombineFunction
                 )
-                .filter(mFilterPredicate)
-                .doAfterNext(new Consumer<BoundData<View, Result>>() {
-                    @Override
-                    public void accept(BoundData<View, Result> viewResultBoundData) throws Exception {
-                        Notification<Result> notification = viewResultBoundData.getData();
-                        if (notification.isOnComplete() || notification.isOnError()) {
-                            try {
-                                mOnTerminate.run();
-                            } catch (Exception e) {
-                                MVPLogger.e(mTag, e.getMessage());
-                            }
-                        }
-                    }
-                });
+                .filter(mFilterPredicate);
 
         // Adds the replaySubject subscription to the CompositeSubscription
         // to be able to dispose the replaySubject from the original flowable